+++ tests/jest/application-error-boundary/ApplicationErrorBoundary.test.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { shallowWithIntl, mountWithIntl } from 'terra-enzyme-intl';
+import { shallowWithIntl, mountWithIntl } from '@cerner/terra-enzyme-intl';
 import ApplicationErrorBoundary from '../../../src/application-error-boundary/ApplicationErrorBoundary';
 import Logger from '../../../src/utils/logger';
 
@@ -8,7 +8,7 @@ describe('ApplicationErrorBoundary', () => {
     it('should render with minimal props', () => {
       const wrapper = shallowWithIntl((
         <ApplicationErrorBoundary />
-      ));
+      )).dive();
 
       expect(wrapper).toMatchSnapshot();
     });
@@ -18,7 +18,7 @@ describe('ApplicationErrorBoundary', () => {
         <ApplicationErrorBoundary>
           <div>Test child</div>
         </ApplicationErrorBoundary>
-      ));
+      )).dive();
 
       expect(wrapper).toMatchSnapshot();
     });
@@ -43,7 +43,7 @@ describe('ApplicationErrorBoundary', () => {
       /**
        * After rendering the component again, the error view should no longer be rendered.
        */
-      wrapper.instance().forceUpdate();
+      wrapper.setProps({});
       expect(wrapper).toMatchSnapshot();
       expect(spy).toHaveBeenCalledTimes(1);
       spy.mockRestore();