+++ tests/jest/application-navigation/terra-application-navigation/tabs/Tabs.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 /* eslint-disable-next-line import/no-extraneous-dependencies */
-import { mountWithIntl } from 'terra-enzyme-intl';
+import { mountWithIntl } from '@cerner/terra-enzyme-intl';
 import Tabs from '../../../../../src/application-navigation/terra-application-navigation/tabs/_Tabs';
 
 describe('Tabs', () => {