+++ tests/jest/application-navigation/terra-application-navigation/utility-menu/UtilityMenu.test.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 /* eslint-disable-next-line import/no-extraneous-dependencies */
-import { mountWithIntl } from 'terra-enzyme-intl';
+import { mountWithIntl } from '@cerner/terra-enzyme-intl';
 import UtilityMenu from '../../../../../src/application-navigation/terra-application-navigation/utility-menu/_UtilityMenu';
 
 describe('UtilityMenu', () => {
   it('should render default element', () => {
     const mountComponent = mountWithIntl(
-      <UtilityMenu.WrappedComponent />,
+      <UtilityMenu />,
     );
 
     expect(mountComponent).toMatchSnapshot();
@@ -14,7 +14,7 @@ describe('UtilityMenu', () => {
 
   it('should render with function callbacks', () => {
     const mountComponent = mountWithIntl(
-      <UtilityMenu.WrappedComponent
+      <UtilityMenu
         hero={<div>my test hero</div>}
         userConfig={{
           name: 'user-name',
@@ -37,7 +37,7 @@ describe('UtilityMenu', () => {
     const mockCallBack = jest.fn();
 
     const mountComponent = mountWithIntl(
-      <UtilityMenu.WrappedComponent
+      <UtilityMenu
         onSelectLogout={mockCallBack}
       />,
     );